modularized methods into DBAuth. toast bug.
